fix(compiler-spec): extract instruction spec derived types
