+++ packages/compiler-spec/src/index.ts
@@ -7,6 +7,7 @@ export * from './diagnostics';
 export * from './errors';
 export * from './functionTypes';
 export * from './instructionSpecs';
+export * from './instructionSpecTypes';
 export * from './instructions';
 export * from './macros';
 export * from './memory';
